services/verification-context/domain/use-cases/src/main/java/uk/co/idv/domain/usecases/verificationcontext/sequences/StubbedSequenceLoader.java
package uk.co.idv.domain.usecases.verificationcontext.sequences;

import uk.co.idv.domain.entities.identity.alias.Alias;
import uk.co.idv.domain.entities.mobiledevice.MobileDevice;
import uk.co.idv.domain.entities.verificationcontext.StubVerificationSequencesEligible;
import uk.co.idv.domain.entities.verificationcontext.StubVerificationSequencesIneligible;
import uk.co.idv.domain.entities.verificationcontext.VerificationSequences;
import uk.co.idv.domain.entities.verificationcontext.method.VerificationMethod;
import uk.co.idv.domain.usecases.verificationcontext.LoadSequencesRequest;
import uk.co.idv.domain.usecases.verificationcontext.SequenceLoader;

import java.util.Collection;

public class StubbedSequenceLoader implements SequenceLoader {

    @Override
    public VerificationSequences loadSequences(final LoadSequencesRequest request) {
        final Alias providedAlias = request.getProvidedAlias();
        if (valueEndsWithNine(providedAlias)) {
            return new StubVerificationSequencesIneligible();
        }
        return new StubVerificationSequencesEligible();
    }

    private boolean valueEndsWithNine(final Alias alias) {
        return alias.getValue().endsWith("9");
    }

}
